ref: move withoutSave function to EncryptSubService
