+++ src/app/modules/submission/encrypt-submission/encrypt-submission.service.ts
@@ -31,7 +31,10 @@ import {
   SubmissionNotFoundError,
 } from '../submission.errors'
 
-import { AttachmentMetadata } from './encrypt-submission.types'
+import {
+  AttachmentMetadata,
+  SaveEncryptSubmissionParams,
+} from './encrypt-submission.types'
 
 const logger = createLoggerWithLabel(module)
 const EncryptSubmissionModel = getEncryptSubmissionModel(mongoose)
@@ -347,4 +350,31 @@ export const checkFormIsEncryptMode = (
   return isFormEncryptMode(form)
     ? ok(form)
     : err(new ResponseModeError(ResponseMode.Encrypt, form.responseMode))
+}
+
+/**
+ * Creates an encrypted submission without saving it to the database.
+ * @param form Document of the form being submitted
+ * @param encryptedContent Encrypted content of submission
+ * @param version Encryption version
+ * @param attachmentMetadata
+ * @param verifiedContent Verified content included in submission, e.g. SingPass ID
+ * @returns Encrypted submission document which has not been saved to database
+ */
+export const createEncryptSubmissionWithoutSave = ({
+  form,
+  encryptedContent,
+  version,
+  attachmentMetadata,
+  verifiedContent,
+}: SaveEncryptSubmissionParams): IEncryptedSubmissionSchema => {
+  return new EncryptSubmissionModel({
+    form: form._id,
+    authType: form.authType,
+    myInfoFields: form.getUniqueMyInfoAttrs(),
+    encryptedContent,
+    verifiedContent,
+    attachmentMetadata,
+    version,
+  })
 }
